+++ src/useDebouncedFn.spec.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import { renderHook, cleanup as cleanupHooks } from '@testing-library/react-hooks';
 import { render, cleanup as cleanupReact } from '@testing-library/react';
-import useDebouncedCallback from './useDebouncedCallback';
+import useDebouncedFn from './useDebouncedFn';
 import promiseDelay from '../test/utils/promiseDelay';
 
 
-describe('useDebouncedCallback', () => {
+describe('useDebouncedFn', () => {
   beforeEach(() => {
     cleanupReact();
     cleanupHooks();
@@ -13,13 +13,13 @@ describe('useDebouncedCallback', () => {
   });
 
   it('should be an arrow function', () => {
-    expect(useDebouncedCallback).to.be.a('function');
-    expect(useDebouncedCallback.prototype).to.be.empty;
+    expect(useDebouncedFn).to.be.a('function');
+    expect(useDebouncedFn.prototype).to.be.empty;
   });
 
   it('should return a single function', () => {
     const fn = () => 0;
-    const { result } = renderHook(() => useDebouncedCallback(fn));
+    const { result } = renderHook(() => useDebouncedFn(fn));
 
     expect(result.current).to.be.a('function');
   });
@@ -28,7 +28,7 @@ describe('useDebouncedCallback', () => {
     const spy = sinon.spy();
 
     const TestComponent = () => {
-      const debouncedCallback = useDebouncedCallback(() => {
+      const debouncedCallback = useDebouncedFn(() => {
         spy();
       }, 250);