+++ src/useMediaQuery.js
@@ -8,21 +8,6 @@ import createCbSetterErrorProxy from './utils/createCbSetterErrorProxy';
  * It also monitor the document changes to detect when it matches or stops matching the media query.<br />
  * Returns the validity state of the given media query.
  *
- * ### Usage
- *
- * ```jsx harmony
- * const MediaQueryReporter = () => {
- *   const isTablet = useMediaQuery('(max-width: 48rem)');
- *   const isDesktop = useMediaQuery('(min-width: 48rem)');
- *
- *   return (
- *     <div style={compStyle}>
- *       <p>Tablet view? {isTablet ? 'yes' : 'no'}</p>
- *       <p>Desktop view? {isDesktop ? 'yes' : 'no'}</p>
- *     </div>
- *    );
- * };
- * ```
  */
 const useMediaQuery = (mediaQuery) => {
   if (!('matchMedia' in window)) return createCbSetterErrorProxy('matchMedia is not supported');