+++ src/useMouseEvents.js
@@ -46,61 +46,6 @@ const assignMouseEventOnMount = (targetRef, handlerRef, eventName) => {
  * lose the React SyntheticEvent performance boost.<br />
  * If you were doing something like the following:
  *
- * ```jsx harmony
- * const MyComponent = (props) => {
- *  const { myCallback } = props;
- *
- *  return <div onMouseDown={myCallback} />
- * }
- * ```
- *
- * **Please keep doing it**!
- *
- * ### Target ref usage:
- *
- * ```jsx harmony
- * const MyComponent = () => {
- *   const ref = useRef();
- *   const { onMouseMove } = useMouseEvents(ref);
- *   const [coordinates, setCoordinates] = useState([0, 0]);
- *
- *   // demo purposes only, in real-life scenario use onMouseMove prop instead
- *   onMouseMove((event) => {
- *     const nextCoords = [event.clientX, event.clientY];
- *     setCoordinates(nextCoords);
- *   });
- *
- *   return (
- *     <div ref={ref}>
- *       The current mouse coordinates within this div are:
- *       <p>x:{coordinates[0]} y:{coordinates[1]}</p>
- *     </div>
- *   );
- * }
- * ```
- * <br />
- *
- * ### Global events usage:
- *
- * ```jsx harmony
- * const MyComponent = () => {
- *   const [coordinates, setCoordinates] = useState([0, 0]);
- *   const { onMouseMove } = useMouseEvents();
- *
- *   // demo purposes only, in real-life scenario use onMouseMove prop instead
- *   onMouseMove((event) => {
- *     const nextCoords = [event.clientX, event.clientY];
- *     setCoordinates(nextCoords);
- *   });
- *
- *   return (
- *     <div style={style}>
- *       The current mouse coordinates within the document are:
- *       <p>x:{coordinates[0]} y:{coordinates[1]}</p>
- *     </div>
- *   );
- *};
- * ```
  */
 const useMouseEvents = (targetRef = null) => {
   const [onMouseDownHandler, setOnMouseDown] = useCallbackRef();